backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0
tinydb==4.8.0
chromadb==0.5.20
sentence-transformers==3.2.0
google-generativeai==0.8.3
langchain-text-splitters==0.3.0
httpx==0.27.0
python-multipart==0.0.12
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
# Production database drivers (used when ENV=prod)
pymongo==4.8.0
psycopg2-binary==2.9.9
